+++ src/java/org/infoglue/cms/util/NotificationMessage.java
@@ -100,7 +100,7 @@ public class NotificationMessage
 	{
 		switch (transactionType.intValue())
 		{
-			case (int) (TRANS_CREATE):
+			case (TRANS_CREATE):
 				return TRANS_CREATE_TEXT;
 			case (TRANS_DELETE):
 				return TRANS_DELETE_TEXT;
